Update patch set 1
Patch Set 1:

Can we think of another way to do this?  

Is it problematic if the USB config jumps through a few items?  If not, then we could leave this as is and make sure the code to read from persist.sys.usb.<OemOverride mode>.config happens after this.

Alternatively, if we're in OemOverride mode, can we clear the persistent configuration such that this never runs?

ro.bootmode is a little bit tricky, and matching to 'unknown' is not going to work well.  Some platforms use 'normal' here, others leave is blank, which results in 'unknown', and there is also recovery mode that we need to handle.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>