Update patch set 1
Patch Set 1:

Reading first has highest priority.

From fs_config perspective, I figured this was most accurate if non-orthogonal content is introduced. But in the comments in fs_config.c and in https://android-review.googlesource.com/#/c/355712 we actually ensure that the content is orthogonal through the filter. We do not want /vendor/etc/fs_config_dirs to affect content that lands in system.img.

So priority does not in fact matter, it is merely stated so that it is clear what is going to happen if someone introduces interference.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>