Update patch set 2
Patch Set 2:

> Original patch was voiding breaking current devices. But it was
 > later split into 2 patches. This one is handling the current
 > devices
 > 
 > https://android-review.googlesource.com/#/c/356334/
 > 
 > From what I understand all slot/suffix reported by bootloader
 > should not have the "_" (you can add that in specific code like in
 > fs_mgr internal variables when needed).
 > 
 > With these 2 patches, bootloaders get updated. Since, without this
 > change if a bootloader gets updated to pass android.slot_suffix=a,
 > it won't work. So, for now "_a" and "a" are both valid cases, later
 > we can remove the "_a" support (reverting change 356334).

According to the current spec, it was correct before. When a variable refers to the slot name, it is 'a' or 'b', but when it refers to the suffix, it is the full '_a' or '_b' suffix. Changing this without changing the variable name is confusing and inconsistent.

Patch-set: 2
Reviewer: Daniel Rosenberg <1045980@85c56323-6fa9-3386-8a01-6480fb634889>