Update patch set 2
Patch Set 2: Code-Review+2

This change itself is fine, but is it possible to timeout the fsck?  It should be, the parent process would just need to kill the fsck process and continue after a given timer expires, right?

The 'delay' is sort of meaningless if fsck takes some time to run.  The delay should encompass the entire reboot process (and fall back to a dirty shutdown) and in the case of delay = 0, we should skip the fsck.

Thoughts?

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2