Update patch set 1
Patch Set 1:

(1 comment)

Do we plan to specify ro.build.shutdown_timeout in any of a product makefile or are we going to assume the normal case is to fall back to the defaults here?  I ask, since it'd be a bit cleaner if in a make file we made this determination rather than at runtime.  

It's strange the way that this is now, as we have two separate defaults for user/userdebug vs eng but both are overridden if there's a property override in the makefiles, so if a device sets a special timeout for itself, then its eng builds will also take that long.

I guess it may be better to have the eng override always override the property?  Even that isn't great, but I can't immediately think of a better way to just override this for eng builds.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>