Update patch set 1
Patch Set 1: Verified+1 Presubmit-Ready+1

In the past, any changes to this code were rejected. Adding QC_RIL may be grounds for this CLs dismissal, as it is considered _better_ to fix the source of the logging.

I could not measure a statistically significant performance difference with this adjustment.

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1