Update patch set 2
Patch Set 2: -Verified

a lot of crash happens when processes are killed. that we should not collect stack as that is not normal order. we need at least one for vendor components. Maybe only keep shutdown_critical for animation and stuffs that should be kept. All others can be just hard coded as there are not that many of them.

Patch-set: 2
Reviewer: Keun-young Park <1010465@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0
Label: Presubmit-Ready=0