Update patch set 1
Patch Set 1:

> a) assigning to salyzyn@android.com is asking for trouble. Please
 > used salyzyn@google.com
 > b) jhawkins@google.com is the defacto maintainer, added him.
 > c) beware of 32-bit tests running on 64 bit platforms, libraries
 > like this will be in /system/lib64/libmetricslogger.so because that
 > is all that is to be delivered, so a 32-bit test will not find
 > /system/lib/libmetricslogger.so
 > 
 > We have the same problem on go/apct for the logcat-unit-test, as
 > liblogcat.so is only supplied as a 64-bit library. See b/34454772
 > for some details of the struggles.

Thanks! I was able to push ./angler/data/nativetest64/logcat-unit-tests/logcat-unit-tests over to the device and it's running fine.

It seems to be a build problem for the test, maybe we should by default save the 64 bit binary to testcases folder for testing.

Patch-set: 1
Reviewer: Dan Shi <1088076@85c56323-6fa9-3386-8a01-6480fb634889>