Update patch set 5
Patch Set 5:

> a lot of crash happens when processes are killed. that we should not collect stack as that is not normal order.

Do we know why this is happening or how many is 'a lot'?  We lose the point of sending SIGTERM if the services are not shutting down gracefully.

One thing I was thinking overnight - iterating with ForEachService() may not be the best idea.  There's no guaranteed ordering there, but it's defacto ordered by how init parses the individual .rc files, which is almost surely not the order in which we want to shut down services.  I can understand why we have crashes if for example hwservicemanager is shut down before any of the HAL services or if servicemanager is shut down before any of the framework services.  

May it be a good idea for init to keep track of the order in which it starts its services and kill them in the reverse order?  Could this help with the crash issues?  This could help stop the special cases too, as logd is started early in the boot process and tombstoned late.

In general, I meant only to special case vold here, as there is already an explicit special case for it due to volume shutdown.  If we really do need the rest of the special cases, I was agreeing with Wei's comment that they should be separate classes.  It'll disturb the init language the least.

Patch-set: 5