Update patch set 1
Patch Set 1:

We debated back on forth on this a bit.  I was in favor of the hardcoding since I really hope that the number of 'critical' services is reasonably low.  I hope other thought to be critical services can be understood to be not actually critical.

We also started supporting multiple groups, so we had thought of using a 'shutdown_critical' group or multiple new service options, but I think your method of a 'shutdown' option that takes an argument is the cleanest if we do go forward with this.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>