Update patch set 2
Patch Set 2:

The (a?) bug is what I described.  killProcessGroupOnce() only calculates 'processes' once, on the first pass through the while loop in killProcessGroup().  As long as there are any processes in the group, regardless of if they've been killed, 'processes' will be >0 and therefore we return -1.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>