Update patch set 1
Patch Set 1: Code-Review+1

Lgtm!

I think the caller is responsible to print the error log when needed.
Because in this function, we won't know the severity.

e.g., in a non-A/B device, there will be, but it's not critical.
Even A/B device might output either of them.
 init: [libfs_mgr]Error finding 'slot' in device tree
 init: [libfs_mgr]Error finding 'slot_suffix' in device tree

Patch-set: 1
Reviewer: Bowgo Tsai <1080813@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1