Update patch set 2
Patch Set 2:

> Patch Set 2: Presubmit-Ready+1
> 
> Ok, this is adding more technical debt. Pick one:
> 
> A) fix all code to recognize libsystem, a recently created header-only library?
> B) merge libsystem into libutils and poof the technical debt goes away.

It's not that simple. This isn't being added just to export libsystem headers everywhere. This is being added because libutils headers are using libsystem headers. So, a client of libutils needs to either import libsystem headers itself, or it should be considered an implementation detail of libutils (which is what is being done here by exporting it).

Patch-set: 2
Reviewer: Steven Moreland <1120458@85c56323-6fa9-3386-8a01-6480fb634889>