Update patch set 1
Patch Set 1:

(1 comment)

I see 5 lines for each exec service now:

[   19.282786] init: starting service 'exec 6 (/system/bin/bootstat -r post_decrypt_time_elapsed)'...
[   19.283445] init: SVC_EXEC pid 502 (uid 0 gid 0+0 context default) started; waiting...
[   19.299309] init: Service 'exec 6 (/system/bin/bootstat -r post_decrypt_time_elapsed)' (pid 502) exited with status 0
[   19.299328] init: SVC_EXEC pid 502 finished...
[   19.299361] init: Wait for exec took 0.0168597 seconds

The bottom three are all redundant, and maybe it's best to collapse them into one.  We're already printing out when a service dies along with its name and error code, so maybe when we have an exec service, we hijack that line to also report how long it took to run?

We can think about doing that for the first two as well, although the first happens before the service starts and the second happens after it has been forked.  Note the 'SVC_EXEC pid ...' format is in theory meant to match with each other, but given neither contain either the name or the time it took to run, it's not super useful.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>