Update patch set 1
Patch Set 1: Presubmit-Ready+1

Admittedly draft quality.

A) Is there a plan for an official keyutils.h and support in bionic?
B) I was tempted to grab the session in the Phase 1 init, so the placement is still arbitrary in my mind

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1