Update patch set 4
Patch Set 4: Verified+1 Presubmit-Ready+1

Only open_memstream.c and fs_config.cpp were built on windows, and the former had the #define of _GNU_SOURCE behind a #if 0 for some regression testing code so not impacted. So we opted to add the definition into the windows build only. (to be removed if the build tools are updated)

Patch-set: 4
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=+1
Label: Presubmit-Ready=+1