Update patch set 4
Patch Set 4:

The build *should* be stable enough for you to say:

mma -j BOARD_VNDK_VERSION=current

to see all the errors that something like this would cause. We're currently trying to get the build in a state that it can build with "the vndk enabled". If we have to fix a couple hundred extra headers over something like this, I don't think it's worth it (yet).

Competing priorities, etc..

Patch-set: 4
Reviewer: Steven Moreland <1120458@85c56323-6fa9-3386-8a01-6480fb634889>