Update patch set 3
Patch Set 3: Code-Review-1

(1 comment)

"this" worked for logd to have access to the device resources and did not appear to report any other issues in the boot up of the device. I can only propose that the error return be noted globally, and if it failed here, retry again later when the mount and activation operations are occurring for the built-in functions that call e4crypt_install_keyring later. There will be only one ;->

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1
Label: Presubmit-Ready=0