Update patch set 1
Patch Set 1:

(1 comment)

This is to pull adb and the test scripts, so that they can be downloaded from the build server (so that we can run these tests on TradeFed). I also plan to add a build target (as in launch control config) for adb_integration_test. 

Are you saying the build team is adding support for this use case? When does it complete?

Patch-set: 1
Reviewer: GuangHui Liu <1156013@85c56323-6fa9-3386-8a01-6480fb634889>