Update patch set 4
Patch Set 4: Code-Review-1

I have had to incorporate vendor, oem and odm in the build system, and we do have risks in oem identified in the fs_config.cpp file. Yes, odm is no where to be found, but if it happens, the cost in engineering time is too great if the expectations are not fully covered.

Patch-set: 4
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=-1