Update patch set 1
Patch Set 1:

I assume you checked that libgraphics is referenced in all those places, or that it is a small effort to add libgraphics to the stragglers.

Q: how many places are touched that will need to add libradio? My worry is that this includes vendor code, requiring them to behave.

Set a topic for this, because if it looks good here, those stragglers in other repositories (within the constraints of aosp) that need changes should be grouped in with this change.

sgtm, but reserve the right to look at what you upload with a face of horror once seen ;-}

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>