Update patch set 1
Patch Set 1:

Don't be confused by --show-example option, it is like a self testing.
I usually use it to trace simpleperf crashes like below:
./static_binary_crash_dump64 ./simpleperf record xxx.
So it runs a command as the child process, and dump callstack when the command crashes because of SIGSEGV.
On Android, if an executable is dynamically linked, the linker will install signal handlers to dump callstack. But for static executables like simpleperf, it is very hard to know where the SIGSEGV happens.
I am not sure whether it is useful to others, so just try it here for review.

Patch-set: 1
Reviewer: Yabin Cui <1056364@85c56323-6fa9-3386-8a01-6480fb634889>