Update patch set 2
Patch Set 2:

Seems right to me given what we're trying to do.  We'll limit ourselves to not being able to have partitions with the same name on multiple block devices, but... that's not a usecase I think we'll ever need to support.

This code already limits us to not having partitions that are a prefix of each other, e.g. "sys" and "system", and that hasn't impacted us yet.

Patch-set: 2
CC: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>