Update patch set 3
Patch Set 3:

This is a PoC for a full blown async property implementation, in case property triggers are not acceptable for some reason.  I'm mostly just throwing it out for comment -- it does have some rough spots that should be ironed out before submit.

I've requested that the bug reporter test the equivalent of this code on our 7.1.x branch and report his results.

Also I still haven't setup a build/test environment for master, so this is untested.  I'll try to do that soon.

Finally, thanks for taking the time to look at this, I really appreciate it.

Patch-set: 3
Reviewer: Tom Marshall <1037023@85c56323-6fa9-3386-8a01-6480fb634889>