Update patch set 3
Patch Set 3:

(1 comment)

Thanks for the implementation.  I haven't had a chance to look into the property trigger idea more, but I'll try to get to it this week.  The biggest issue is going to be performance, because if there is anything else in the Action queue, when restorecon is called (which may be likely), then this is a non-started

Otherwise, I still want to see if we can simplify this a bit.  Perhaps if we had a 2nd variable (like the property trigger idea) to record the successful restorecon, we could skip having to do in the sigchld handler as the subprocess could set other properties without issues with android::base::SetProperty().  We could use that property set as an action to fork() pending restorecon paths as well.

Patch-set: 3
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>