Update patch set 1
Patch Set 1:

Hi Jayant and Vijay, 

Could you tell me why this is marked as vendor_available? I see no vendor lib (or other vendor_available lib) is using this.

I know that it is in the eligible-vndk-list but I think we might have to remove it from the list since no one is using it.

Patch-set: 1
CC: Jiyong Park <1132673@85c56323-6fa9-3386-8a01-6480fb634889>