Update patch set 3
Patch Set 3:

(2 comments)

I think the aim for this change is good, since it allows us simply poll and create devices without having to go through the coldboot first (which creates the platform device list). This also means the symlinks won't be missed if the device appears on the platform bus at runtime (after boot) and that I doubt happens on devices today.

Overall, LGTM, but want to make sure that the old "keep the platform devices list and traverse device directory only if needed" is also part of this CL instead of getting rid of it.

Patch-set: 3
Reviewer: Sandeep Patil <1108565@85c56323-6fa9-3386-8a01-6480fb634889>