Update patch set 3
Patch Set 3:

(2 comments)

> But want to make sure that the old "keep the platform devices list and traverse device directory only if needed" is also part of this CL instead of getting rid of it.

For speed or why?  I feel like it's cleaner to traverse first than rely on the overall list (plus then we won't need to have a list of all platform devices when we only care of a few, nor process those uevents).

It'll be hard to know when we need to fallback to the traversal anyway.  /sys/devices/soc.0 is a platform device on sailfish, so it's possible that we match that and think we've found our platform device, whereas in reality we're looking for a child off of that.

Patch-set: 3
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>