Update patch set 5
Patch Set 5:

> Regarding code style (type_*_variable): file has to have only one code style. Other functions in this project have "type *variable" style.
> If you want to change codestyle than you should do it in one separate commit.

Normally I'd agree, but that's not what we decided for most of system/core and this file is already mixed.  We're fixing the style as we make changes instead of doing a bulk update.  It makes it super easy to see what code was written Pre/Post moving to C++.  The .clang-format agrees with this and will correct the format once it's run.

Patch-set: 5