Update patch set 1
Patch Set 1:

Is this file really required for every single gtest?  I copied the one from libcutils and did a search and replace, which is a sign that it contains no useful content.  For most gtests, all the test system should need to know is that it is a gtest.

Patch-set: 1
Reviewer: Colin Cross <1002751@85c56323-6fa9-3386-8a01-6480fb634889>