Update patch set 2
Patch Set 2:

For standardization between CTS/APCT and our future plans for atest/generalized suites, yes it is required.

The idea of the configs is it informs the test runner completely on how to run an individual test and any special prep the test author may require can be defined in this file. I do like the idea of autogenerating it for tests that don't have any special requirements so we can try to get that work prioritized soon.

Patch-set: 2
Reviewer: Simran Basi <1083173@85c56323-6fa9-3386-8a01-6480fb634889>