Update patch set 2
Patch Set 2:

This solution is KISS and glosses over having to expose event logs in vendor code. How many other places are these calls? If this is the _only_ place then I am _happy_; but if there are more, then some work in liblog (to expose the macros) should be considered.


[TL;DR]

If there were not some small (mainly byte ordering, considered a non-issue) issues with the LOG_EVENT_* macros, they expose implementation details, and they can not handle multiple arguments or the general case; I would not have any issues exposing them in liblog. The main advantage of them is that the LOG_EVENT_* macros will work over a wide range of releases, whereas the event log list handlers which abstract the implementation details were added in M. They are in extensive use via JNI and in all places where more complex event message are composed.

There has been a case made that a vendor version of libsysutils should not be sending event messages as they are part of services where we do not care for the reporting. We do not care for vendor code spamming the events log...

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>