Update patch set 2
Patch Set 2:

And the selinux ones are to be replaced with calls to the netlink socket into the kernel so that they can be reported correctly, so that one in time will be moot.

As noted, there is nothing secret about these macros, the preference _I_ have is to remove examples of their use (deprecate them). enh@'s team is the lead on the logger so what they say goes.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>