Update patch set 2
Patch Set 2:

More basic Q than that: do we _want_ the Netlink failure logging for ueventd (and thus for the vendor brethren that may do the same)

If not, then this change stands as-is.

As to the former: even if you did use the library, the fact ueventd is running AID_ROOT would mean no message would have been produced as it checks for a positive value for uid.

As to latter: ueventd can log to the Android logger near any time it wants, logd is started right _after_ ueventd and effectively starts taking content as soon as init has created the socket (even before 'exec' of logd). events buffer logging like this is certainly doable within the context of the existing code.

If it was to log the message, my hopes you use the C++ interface and the event tag lookup (we don't have an inline one yet for picking up the log tags yet):

{
  static EventTagMap* map;
  if (!map) map = android_openEventTagMap(nulptr);
  static uint32 tag = android_lookupEventTagNum(map, "exp_det_netlink_failure", "(uid|1)", ANDROID_LOG_INFO);
  android_log_event_list(tag) << (int32_t)getuid() << LOG_ID_EVENTS:
}

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>