Update patch set 2
Patch Set 2:

But why is this useful?  The VNDK builds are going to have tons of warnings from the current legacy inclusions of this header.  I doubt that anyone is going to even pay attention if new warnings show up.  If we come up with a concrete plan to deprecate this header, then sure let's add the warning as we make the changes, but until then it'll just add more code and main pain for everyone.

Patch-set: 2
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>