Update patch set 1
Patch Set 1: Presubmit-Ready+1

Thanks for the quick response! I didn't realize there was a prior attempt. I am seeing that your earlier CL also had a test of this feature. Is there a way to test that compilation __fails__ for a certain piece of code, or at least to test that a warning is generated by a compiler?

Patch-set: 1
Reviewer: Siarhei Vishniakou <1149384@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1