Update patch set 1
Patch Set 1:

Understood.  I still don't see why this is init's bug to fix.  The goal of the emulation should be to get as close to the final system as possible.  If there's an issue because the emulator doesn't have a partition table, then the solution here is simple: the emulator needs to use a partition table and expose the system (and later vendor, etc) partitions the same as a real device would.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>