Update patch set 2
Patch Set 2:

I checked that the warning is generated only if -std=c++11 is passed, so Clang is doing the right thing.  TH failed because these headers can also be used both in non-c++11 modules (where nullptr literal is not defined).

I think the proper fix would be to disable -Wzero-as-null-pointer-constant via pragmas in these headers so we don't have to disable it globally.

Unfortunately, this is a new warning and we cannot submit it before a prebuilts update.

Patch-set: 2
Reviewer: Stephen Hines <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Reviewer: Pirama Arumuga Nainar <1058860@85c56323-6fa9-3386-8a01-6480fb634889>