Update patch set 1
Patch Set 1:

(1 comment)

Even with the inlining (and likely due to it), the init executable is smaller with this CL, so there's no concern of over using that.

Eventually this won't be a singleton.  This is going to help achieve that goal.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>