Update patch set 4
Patch Set 4:

(5 comments)

I need to update the commit message and the write up, but this seems to make sense now.

I have a test that ensures that this is at worst a move construction, so it shouldn't be much of a regression over passing the parameter on the command line.

Patch-set: 4
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>