Update patch set 1
Patch Set 1:

Specifically setsockcreatecon() isn't in libselinux.so but it's in libselinux.a.  I'm pretty sure I've hit similar issues before when trying to write other tests, so I figure this is the right approach.

Otherwise, I could add setsockcreatecon() to libselinux.so?  But it seems like that's not the right solution, as it looks like we want to keep as few symbols in there as possible.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>