Update patch set 1
Patch Set 1:

> This is a system health issue. Gaining ota and factor_reset
 > detection was a nice-to-have, canonical boot reason seems ideally
 > suited here in my mind. But maybe we need to add (parts of? whole?)
 > bootstat as part of the system health bundle?

I'm OK with the proposal.  I intend for the command to continue existing, as the data reported will always be valuable (long term to catch regressions or outlier devices), so it should be safe to have the functionality here.

I'll be watching for your second patch set for followup review.

Patch-set: 1
Reviewer: James Hawkins <1081159@85c56323-6fa9-3386-8a01-6480fb634889>