Update patch set 1
Patch Set 1:

You need to do a rebase, I made a small change to some of the code you are modifying.

And yes, that backtrace tests are incredibly slow. It's on my list of things to fix, along with fixing all of the backtrace test code.

Patch-set: 1
Reviewer: Christopher Ferris <1019050@85c56323-6fa9-3386-8a01-6480fb634889>