Update patch set 3
Patch Set 3:

(1 comment)

> logd-reinit is called by Setup application when changes are made to the log buffer size.

The Setup application as in the Setup app way up in Java?  Can you point me to the code?

It only needs to do the logbuf reinitialization, not the android::ReReadEventLogTags(); which is the main part I'm trying to avoid, right?

Does the logBuf reinitialization need to happen in a separate thread?  Are there locks preventing it from being done in CommandListener.cpp directly without the signal and semaphore?

Patch-set: 3
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>