Update patch set 1
Patch Set 1:

This is a brutal hack, but so is a lot of how we initiate shutdown.

Alternatively, I'm thinking I could have an epoll_handler for this to save on the init globals, but that is more changes and it'd likely require us to handle more than one of the epoll events at a single time, which would be a change needing more testing.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>