Update patch set 1
Patch Set 1:

So the one major design consideration here is whether or not Subcontext::Execute() should be synchronous or asynchronous.  

In the latter case, we'd be able to handle restarting services and property setting while a vendor command is running, but it comes with refactoring Action::ExecuteOneCommand() and ActionManager::ExecuteOneCommand().  

That said, we don't handle those in the background of commands currently, and the added communication over this socket shouldn't take so long that we'd get behind in processing those events, so I think we're okay.  Additionally, at some point, property handling will be removed from init, so that's one less reason to add the complexity to make it asynchronous.

This is definitely WIP though; in particular, I need a mechanism for recovery if a subcontext init crashes.  We'll need to define the SELinux context for 'vendor_init' as well before this can boot.  Coincidentally the 'init' context isn't enough, since init doesn't have permissions to write to the property socket.  Otherwise setting SELinux to permissive can help test the mechanism in general.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>