Update patch set 2
Patch Set 2: Presubmit-Ready+1

In the context of the bug we need to contact the batterystat service, and reading comment from a pipe connected via a binder call would require a timeout.

Outside the bug, this additional interface can be rolled over the entire tree anywhere we see C++ code that does poll timeout and read in combination. The other 'side' can be a pipe, or socket. The need is there to remove open coding of this.

For instance, I used dumpsys as the original code I copy-n-pasted from (well, because the private branch'd experimental code is in-effect doing a dumpsys batterystats), and you just found a bug with it(!). frameworks/cmd/dumpsys/dumpsys.cpp could very well use libbase and become hardier and simpler in the process.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1