Update patch set 2
Patch Set 2:

(3 comments)

I think this check ideally belongs in a linter, but until we start running one of those as a part of regular builds of Android, this is probably the best we can do.

LGTM with a few comment nits. Thanks!

Patch-set: 2
Reviewer: George Burgess IV <1096946@85c56323-6fa9-3386-8a01-6480fb634889>