Update patch set 8
Patch Set 8:

max_size is an _optional_ parameter, so I do not see it as annoying. No one has to use it, or specify it, and the availability of it makes a _big_ difference to the number of places this enhancement to the API can be used.

Agreed on the 5 second assumption, I was nervous about that, if removed we can merely state that a timeout of 0 only ensures that the first 1K or max_size, whichever is less, is guaranteed to be read, if ready to pick up. Or we could let it always try to fulfill max_size (if not defaulted to SIZE_MAX) though allowing an indefinite stretch of timeout while content continues flowing.

Patch-set: 8
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Verified=0