Update patch set 9
Patch Set 9:

(2 comments)

The hero locations for this activity reside in dumpstate and anyone who wants to replicate dumpstate (ie: service->dump() calls return a pipe fd). Discussions with tomcherry@ has us considering making something like this as part of libdumpstate instead of libbase.

We have evaluated that even though this can work for adb, it costs one syscall to requests a poll from a file descriptor that may not be a socket/pipe. The windows code in adb tracks all open file descriptors with a class, so it knows it is dealing with a socket and does not waste the call on a file. Best to leave that one alone, and in doing so, we remove the need to provide windows support (here or in libdumpstate).

Bootanimation was not so heroic because of FUD associated with alignment and the issue with discerning error or timeout. It works, but makes all of us nervous.

debuggerd is a 'heroic', but we can walk away from it as it was already KISS / correct, or borrow from libdumpstate instead.

Patch-set: 9
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0