Update patch set 4
Patch Set 4:

Code looks good to me.

A question I should have been asking yesterday: Is there a reason why we're not letting surfaceflinger (or another graphics process) handle this?  Presumably SF can do the blanking of the contents and the 0'ing of the backlight on the suspend/hibernate path such that the "does not mean the content is not visible" issue goes away.  

SF could/should also be responsible for unblanking the screen and restoring the backlight.  As it stands, we're almost certain to have a glitch on the panel when this code runs, as there will be some time between when the backlight comes back on and the next frame is rendered.  I'm sure, even if it isn't now, that removing this glitch will be a priority in the future for quality/polish reasons.

Patch-set: 4
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>