Update patch set 4
Patch Set 4:

Do you want the following to be added/paraphrased to the commit message (I would expect so...)

This is the lowest (hardware/operating system) implementation that preserves the boot reason along with the associated kernel actions. It is the responsibility of the caller to make all the necessary arrangements (as discussed in 'init: support suspend and hibernation' CL commit message, and in the associated bugs). The caller is expected to be the PowerManager service implement a broadcast to the services that would be impacted by power or timing issues with a forced suspend or hibernate and would be part of the API implementations at the PowerManager level.

I do _not_ expect init to do any direct calls to SF or other services, I do not expect init to establish any binder communications to issue a broadcast.

If the panel was been appropriately cleared _and_ blanked, then the change here does nothing of value, setting a panel at 0 brightness to  0 before and after return. This change is all about failsafe.

An no, the OS does _not_ turn the backlight to 0 on suspend, it can be on a device basis of course. The use case for casual auto suspend when all wakelocks are released to permit reading (books) or for the watch face to be run at lowest brightness level while the system is put into a low power suspend to RAM state. On the Hikey, suspend and hibernate result in frozen content on the HDMI display; no actions could be performed at the init and OS layer to deal with that. As such we can see a use case for Android TV (letting TV processing continue to HDMI while android suspends/hibernates).

Of course, if forced suspend (or hibernate) are eventually used for the screen-on use cases, we expect adjustments to this, either in ro.support.<type> values or in the third argument at runtime. The use cases right now (Android Auto) do not include such.

Patch-set: 4
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0