Update patch set 2
Patch Set 2:

Nothing compelling, as this is an example for spaik@ to test, and any changes here to refine it may, or may not, go into init, but they certainly will go somewhere. I consider this a prototype only that needs a strong review (hence why you are included)

1) suspend and hibernate for _all_ devices need to be managed by the PowerManagementService. It will not be implemented correctly because of thundering-herd concerns, and complete distrust for applications to support the broadcast appropriately. Requirement changed to only support forced suspend and forced hibernate and require all disrupted services to handle any issues they have with CPU asynchronously suspending for long periods of time.
2) The only supported APIs for power/reboot control communication to the kernel and Canonical Boot Reason are libcutils android_reboot and its implementation detail to write to sys.powerctl. We need to take as many atomic details as close to the suspend request to the kernel as we can because any delays up or down introduce more race conditions.

writes to sys.powerctl property are not _positive_ as the actions occur _after_ the write has completed, hence why I had to move the code into init to prototype the actions.

Moving _all_ of this forced-suspend actions out of init and into a native power management service, or perhaps the libsuspend library, but with a default and vendor HAL is the preference.

Patch-set: 2
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>