Update patch set 16
Patch Set 16:

(2 comments)

Just the one `if` in InvokeFunc() right?  My biggest problem with getting rid of it, is that we still need an if for execute_in_subcontext_, since many commands don't actually get sent to the subprocess, so we'd end up with something like

BuiltinExecutor::Execute(...) {
    return RunBuiltinFunction(...);
}

SubcontextExecutor::Execute(...) {
    if (execute_in_subcontext) {
        subcontext_->Execute(...);
    } else {
        RunBuiltinFunction(...);
    }
}

among the rest of the boilerplate just to create those classes and an executor interface.  Is that really worth it?

Do we gain much by passing some data object vs having the two parameters to builtins?  It'd be simple if we didn't need to expand args before we use them, since then we can just pass `Command`.

Unless... we assume that we only reference the arguments in `Command` more than once and therefore it'd be fine to just have an accessor that provides access to args(x) and does the expand before returning?  Let me look at doing this.

Patch-set: 16
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>