Update patch set 16
Patch Set 16:

> Unless... we assume that we only reference the arguments in `Command` more than once and therefore it'd be fine to just have an accessor that provides access to args(x) and does the expand before returning?  Let me look at doing this.

Nevermind, this is a bad idea from the start.  I don't recreate a `Command` object after sending a message to the subcontext and we proactively return a failure instead of calling the function if any of its arguments cannot be expanded.  Making this change would remove that proactive error.

I can still build a data class if that'd be prettier, but it'd literally just be a data class.

Patch-set: 16
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>