Update patch set 15
Patch Set 15:

=== Finished presubmit run: L84600000107340626 ===

Status: https://android-build.googleplex.com/builds/where-is-my-cl/android-review/485308?revision=15




NOTE: This comment applies to an older patchset.

TreeHugger finished with: 17 passed, 1 failed.





WARNING: The tip-of-tree(4368301) for git_master/errorprone_fastbuild3e_linux may be broken.
The presubmit build passes with the last-known-good-build(4368206), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.



WARNING: The tip-of-tree(4368301) for git_master/gce_x86_phone-userdebug_fastbuild3c_linux may be broken.
The presubmit build passes with the last-known-good-build(4368141), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.







FAIL: Tests failed for git_master/marlin-userdebug/4368088 (LKGB). <apct/ota/recovery> 148 ran, 1 failed, Attempt 5ccbc1ca-5347-412a-9a30-846eede79fd9: URI: http://sponge.corp.google.com/invocation?tab=Test+Cases&show=FAILED&show=INTERNAL_ERROR&id=e5a91e64-7f3a-46f1-bf1d-6381722ad12a ; Check recent test results here: http://go/atp-test-status?f=branch:in:git_master&f=build_target:in:marlin-userdebug&f=test_name:in:apct/ota/recovery
WARNING: The tip-of-tree(4368301) for git_master/marlin-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4368088), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4368301) for git_master/marlin-userdebug_fastbuild_linux may be broken.
The presubmit build passes with the last-known-good-build(4368088), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4368301) for git_master/sailfish-userdebug_fastbuild3_linux may be broken.
The presubmit build passes with the last-known-good-build(4368055), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.




WARNING: The tip-of-tree(4368301) for git_master/taimen-userdebug may be broken.
The presubmit build passes with the last-known-good-build(4368141), so
we assume a failure in ToT is a pre-existing error. Tests will use
the LKGB to try to workaround the pre-existing error.





Have a question? Found a bug? http://go/treehugger-troubleshoot

Patch-set: 15
Tag: autogenerated:TreeHugger:Verify-Stale-Fail