Update patch set 1
Patch Set 1:

Thanks.

Well, you could know that your are _not_ in that state :-) test-for-not-bootloader ...

Again, I have the workaround for test-for-bootloader functionality using fastboot device command. Yes it would be nice if we added (?) the test or wait (?) feature to fastboot command.

Based on your comment we can respin _this_ CL to drop all mention of bootloader state support in adb (I will push the latest, but still incomplete WIP, just for history sake in case the future would like to reverse this course).

Patch-set: 1
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>