Update patch set 1
Patch Set 1: Presubmit-Ready+1

I know properties aren't great, but I figure this would be the easiest way to flag this on/off while the SEPolicy is developed and this is enabled across various devices.  Plan would be to remove it as soon as that's done.

Patch-set: 1
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1