Update patch set 4
Patch Set 4:

This supports a shutdown sequence of:

adb shell reboot -p
adb wait-for-not-device

For forced suspend or hibernate proposed for libsuspend, adb wait-for-not-device can help with sequencing a test script as well, and why it was an interesting addition.

adb reboot -p could have a "wait-for-not-device" operation rather than an indefinite pause. But not sure if that is worthwhile or practical to be built-in.

Patch-set: 4
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0