Update patch set 3
Patch Set 3:

I understand that, but this change doesn't impact that, it just changes the mechanism through which it works.  I'm wondering if there was a reason to completely skip verity on eng vs use the same bit as `disable-verity` as this change seeks to do.

Patch-set: 3
Reviewer: Tom Cherry <1064128@85c56323-6fa9-3386-8a01-6480fb634889>