Update patch set 3
Patch Set 3:

(1 comment)

Kernel command line android.bootreason= is how the bootloader communicates to ro.boot.bootreason. Bootloaders must _always_ provide this value, and in a format that we mandate. Failure to provide one is a CTS failure.

We do not want to _fix_ the platform to allow ro.boot.bootreason parameter to be altered/injected on any (especially "user") builds, could form an attack vector since partner code may use it to communicate to their HALs or services and the code paths may themselves be vulnerabilities. Any means we open up to allow this to happen on the 'platform' would be bad IMHO.

I prefer this slippery means of injecting ro.boot.bootreason values to test bootstat, for development builds, and combined with errant bootloaders, only. Anything else would need a security review of all components affected.

We have bootstat (at a later point) filing in ro.boot.bootreason if it is not set to make sure it is locked down should the bootloader fail to set it. Bootstat does record and propagate if the bootloader fails to set this value on the population of all devices to evaluate the scale of the problem.

Patch-set: 3
Reviewer: Mark Salyzyn <1032276@85c56323-6fa9-3386-8a01-6480fb634889>